Cleaned the classes involved with mobs and pathfinding
