+++ Assets/Resources/AI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1ed2ac00680dcc94a9f5dd171aa48218
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
