+++ Assets/Scenes/Test/MobsTest.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 7e60852a7dcd47f40bdf7a81c7c2946e
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
