--- Assets/Scripts/AStar.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f7e4da86651fd794db2b9c460bbcf6ac
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
