--- Assets/Scripts/Mob.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3b7ff53f9f23ca24e8c918083f6180de
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
