--- Assets/Scripts/Mob/GenericMob.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fcef806299ee4504b8714eba18a2aeb1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
