--- Assets/Scripts/Mob/MobAI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3fef1ab332ec7eb42abae06ac09cd572
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
