--- Assets/Scripts/Mob/Spawnest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 95c21fa31a1e0dc449a638f72582a489
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
