+++ Assets/Scripts/Mobs/AStar.cs
@@ -1,0 +1,9 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using Pathfinding;
+
+public class AStar : MonoBehaviour
+{
+
+}
