+++ Assets/Scripts/Mobs/Crawler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 98b968842b1d0ac428c5dd0f92e97de8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
