+++ Assets/Scripts/Mobs/Flydier.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1797f360297174f48a942301f2e6db10
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
