+++ Assets/Scripts/Mobs/PointToAttack.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2da83c00be321864ba4b1c08d53aa50c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
