+++ Packages/manifest.json
@@ -3,8 +3,8 @@
     "com.unity.collab-proxy": "1.15.4",
     "com.unity.feature.2d": "1.0.0",
     "com.unity.ide.rider": "3.0.7",
-    "com.unity.ide.visualstudio": "2.0.12",
-    "com.unity.ide.vscode": "1.2.4",
+    "com.unity.ide.visualstudio": "2.0.16",
+    "com.unity.ide.vscode": "1.2.5",
     "com.unity.test-framework": "1.1.29",
     "com.unity.textmeshpro": "3.0.6",
     "com.unity.timeline": "1.6.3",
